refactor: remove dead code, break the utilities cycle, fix the drawer width
initPlayerInContainer built a player into a container, which is what the
drawer does, and nothing had called it since the drawer landed. It was
also still emitting Bandcamp's own class names, the collision this
branch spent its time removing. Gone, along with the tracklist builder
and container type only it used, and with injectTrackData, which was
similarly unreferenced.

utilities imported fillFrame from label_view while label_view imported
attachPreviewListeners back from utilities. That cycle meant importing
any helper pulled in the entire player, so background modules carried
content script code and a gap in one mock could take out unrelated
suites. attachPreviewListeners is preview logic rather than a general
utility, so it now sits beside fillFrame. cache_backend no longer
reaches playerLoader, playerDrawer, label_view or nativePlayerBuilder.

Import specifiers were a mix of extensionless and .js for the same
files, which is why the language server treated './utilities' and
'./utilities.js' as separate modules and under-reported references.
Standardised on extensionless.

The drawer sized itself with 33vw and three media queries, so it
resized as the window did. It now takes a single declared width, with
the overlay bound to the same custom property, and yields only when the
window is narrower than the drawer itself.
